requirements.txt
tensorflow==2.8.0
opencv-python==4.5.5.64
numpy==1.21.5
Pillow==9.0.1
matplotlib==3.5.1
scikit-learn==1.0.2